Variables changed from float to double

+++ RTv1/include/rtv1.h
@@ -6,7 +6,7 @@
 /*   By: wwatkins <wwatkins@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/28 10:54:12 by wwatkins          #+#    #+#             */
-/*   Updated: 2016/02/09 07:55:18 by wwatkins         ###   ########.fr       */
+/*   Updated: 2016/02/09 08:12:50 by wwatkins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,11 +66,11 @@ typedef struct	s_cam
 	t_vec3	pos;
 	t_vec3	dir;
 	t_vec3	origin;
-	float	dist;
-	float	w;
-	float	h;
-	float	xi;
-	float	yi;
+	double	dist;
+	double	w;
+	double	h;
+	double	xi;
+	double	yi;
 	int		fov;
 }				t_cam;
 
@@ -175,10 +175,10 @@ void			raytracing_draw(t_env *e);
 **	ray_intersect.c functions
 */
 
-t_obj			*ray_intersect(t_env *e, float *tmin);
-float			ray_intersect_cone(t_env *e, t_obj *obj);
-float			ray_intersect_plane(t_env *e, t_obj *obj);
-float			ray_intersect_sphere(t_env *e, t_obj *obj);
-float			ray_intersect_cylinder(t_env *e, t_obj *obj);
+t_obj			*ray_intersect(t_env *e, double *tmin);
+double			ray_intersect_cone(t_env *e, t_obj *obj);
+double			ray_intersect_plane(t_env *e, t_obj *obj);
+double			ray_intersect_sphere(t_env *e, t_obj *obj);
+double			ray_intersect_cylinder(t_env *e, t_obj *obj);
 
 #endif
